refactor: moves folders to java like package
