+++ src/main/java/com/pauloelienay/slinky/generics/IGenericBusiness.java
@@ -1,0 +1,6 @@
+package com.pauloelienay.slinky.generics;
+
+import java.io.Serializable;
+
+public interface IGenericBusiness<T extends IGenericEntity<T, S>, S extends Serializable> {
+}
