+++ src/main/java/com/pauloelienay/slinky/utils/StringUtils.java
@@ -1,0 +1,22 @@
+package com.pauloelienay.slinky.utils;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static java.util.Objects.isNull;
+
+@SuppressWarnings({"unused"})
+public class StringUtils {
+	public static List<String> toList(String... strings) {
+		return Arrays.stream(strings).collect(Collectors.toList());
+	}
+
+	public static boolean isNullOrBlank(String s) {
+		return isNull(s) || s.isBlank();
+	}
+
+	public static boolean isNonNullOrBlank(String s) {
+		return !isNullOrBlank(s);
+	}
+}
