--- src/main/java/generics/IGenericBusiness.java
@@ -1,6 +1,0 @@
-package generics;
-
-import java.io.Serializable;
-
-public interface IGenericBusiness<T extends IGenericEntity<T, S>, S extends Serializable> {
-}
