--- src/main/java/generics/IGenericEntity.java
@@ -1,14 +1,0 @@
-package generics;
-
-import java.io.Serializable;
-
-import static java.util.Objects.isNull;
-
-@SuppressWarnings({"unused"})
-public interface IGenericEntity<T, S extends Serializable> {
-	S getId();
-
-	default boolean isUpdate() {
-		return isNull(getId());
-	}
-}
