+++ src/test/java/com/pauloelienay/slinky/utils/readonly/EmbeddableObject.java
@@ -1,0 +1,9 @@
+package com.pauloelienay.slinky.utils.readonly;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+@SuppressWarnings({"unused"})
+public class EmbeddableObject {
+	private String name;
+}
