+++ src/test/java/com/pauloelienay/slinky/utils/readonly/ObjectMappedByObjectWithMappedBy.java
@@ -1,0 +1,11 @@
+package com.pauloelienay.slinky.utils.readonly;
+
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+
+public class ObjectMappedByObjectWithMappedBy {
+	@Id
+	Long id;
+	@OneToOne()
+	private ObjectWithMappedBy object;
+}
