+++ src/test/java/com/pauloelienay/slinky/utils/readonly/ObjectWithSerialVersionField.java
@@ -1,0 +1,9 @@
+package com.pauloelienay.slinky.utils.readonly;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@SuppressWarnings({"unused"})
+public class ObjectWithSerialVersionField {
+	private static final long serialVersionUID = 1L;
+}
