+++ src/test/java/com/pauloelienay/slinky/utils/readonly/ObjectWithTransientField.java
@@ -1,0 +1,9 @@
+package com.pauloelienay.slinky.utils.readonly;
+
+import javax.persistence.Transient;
+
+@SuppressWarnings({"unused"})
+public class ObjectWithTransientField {
+	@Transient
+	private String name;
+}
