+++ src/test/java/com/pauloelienay/slinky/utils/readonly/ObjectWithUpdatableField.java
@@ -1,0 +1,12 @@
+package com.pauloelienay.slinky.utils.readonly;
+
+import lombok.AllArgsConstructor;
+
+import javax.persistence.Column;
+
+@SuppressWarnings({"unused"})
+@AllArgsConstructor
+public class ObjectWithUpdatableField {
+	@Column
+	private String name;
+}
