--- src/test/java/utils/readonly/EmbeddableObject.java
@@ -1,9 +1,0 @@
-package utils.readonly;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-@SuppressWarnings({"unused"})
-public class EmbeddableObject {
-	private String name;
-}
