--- src/test/java/utils/readonly/ObjectMappedByObjectWithMappedBy.java
@@ -1,11 +1,0 @@
-package utils.readonly;
-
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-public class ObjectMappedByObjectWithMappedBy {
-	@Id
-	Long id;
-	@OneToOne()
-	private ObjectWithMappedBy object;
-}
