--- src/test/java/utils/readonly/ObjectWithMappedBy.java
@@ -1,14 +1,0 @@
-package utils.readonly;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.OneToOne;
-
-@SuppressWarnings({"unused"})
-@AllArgsConstructor
-@NoArgsConstructor
-public class ObjectWithMappedBy {
-	@OneToOne()
-	ObjectWithUpdatableField object;
-}
