--- src/test/java/utils/readonly/ObjectWithSerialVersionField.java
@@ -1,9 +1,0 @@
-package utils.readonly;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@SuppressWarnings({"unused"})
-public class ObjectWithSerialVersionField {
-	private static final long serialVersionUID = 1L;
-}
