--- src/test/java/utils/readonly/ObjectWithTransientField.java
@@ -1,9 +1,0 @@
-package utils.readonly;
-
-import javax.persistence.Transient;
-
-@SuppressWarnings({"unused"})
-public class ObjectWithTransientField {
-	@Transient
-	private String name;
-}
