--- src/test/java/utils/readonly/ObjectWithUpdatableField.java
@@ -1,12 +1,0 @@
-package utils.readonly;
-
-import lombok.AllArgsConstructor;
-
-import javax.persistence.Column;
-
-@SuppressWarnings({"unused"})
-@AllArgsConstructor
-public class ObjectWithUpdatableField {
-	@Column
-	private String name;
-}
